requirements.txt
biopython
torch
torchvision
scikit-learn
numpy
pandas
matplotlib
seaborn
requests
cutadapt
streamlit